requirements.txt
# UI
streamlit==1.30.0

# LangChain ecosystem (Python 3.12 compatible)
langchain==0.2.12
langchain-community==0.2.11
langgraph==0.0.69

# Gemini (FIXED COMPATIBILITY)
google-generativeai==0.7.2
langchain-google-genai==1.0.8

# Vector DB
weaviate-client==4.5.7

# Graph DB
nebula3-python==3.8.3

# Document processing
pypdf==4.2.0
python-docx==1.1.2
openpyxl==3.1.5

# Utils
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2

# Observability
langsmith
